Handle when sponsors engage in harassing behavior
Currently the code of conduct says sponsors are also subject to the anti-harassment policy, but doesn't state what action will be taken if they breach it (sponsors are not defined as a subset of participants, so as it stands they are not included). To clarify this I've included Sponsors in the 'action' paragraph.