Clarify and strengthen language.
I intended to make the policy clearer on certain details and use more
powerful language. I also made certain parts of the policy focus more
on actions than intent because actions are a more concrete, less
ambiguous, and less abusable type of language to base a policy off
of. I also edited some parts to add the adjective off-topic to the
rules about sexual content because certain topics require discussion
of sexual issues. As well, I switched a few parts to use abuse instead
of harassment because it's shorter and more impactful and also because
of semantic satiation.

I think the biggest change was in:

<p>Harassment includes offensive verbal comments related to gender,
age, sexual orientation, disability, physical appearance, body size,
race, religion, sexual images in public spaces, deliberate
intimidation, stalking, following, harassing photography or recording,
sustained disruption of talks or other events, inappropriate physical
contact, and unwelcome sexual attention.</p>

where I removed a whole list of things that offensive verbal comments
could be about, reordered actions by approximate seriousness and added
a few types of things that harassment could be. I removed the list of
possible offensive topics because I felt that the list could never be
complete and that the actual topic itself is beside the point, the
reason why offensive comments are bad is not because they hurt
people's feelings but because the language is not intended to serve a
constructive purpose but only to hurt people and also because the
comments are off-topic for the conference.