MAY take any action -> COMMIT to take any action
The whole point is that organisers are committed to making spaces safer.

It's easy and vague enough to avoid responsibility with "any action they deem appropriate" just by saying "I did not deem this appropriate". I suggest that we at least take a stance so that people acknowledge problems and do some sort of due diligence.